producer/requirements.txt
kafka-python==2.0.2
pydantic==2.4.2
faker==19.10.0
python-json-logger==2.0.7